+++ src/RequestForm.js
@@ -750,7 +750,6 @@ class RequestForm extends React.Component {
                                   withRef
                                   ref={this.itemBarcodeRef}
                                   onKeyDown={e => this.onKeyDown(e, 'item')}
-
                                 />
                               )}
                             </FormattedMessage>
@@ -800,7 +799,6 @@ class RequestForm extends React.Component {
                             />
                           </span>
                         }
-
                         {multiRequestTypesVisible &&
                           <Field
                             label={<FormattedMessage id="ui-requests.requestType" />}
@@ -933,7 +931,6 @@ class RequestForm extends React.Component {
                               visibleColumns={['active', 'name', 'patronGroup', 'username', 'barcode']}
                               columnMapping={columnMapping}
                             />
-
                           </Col>
                           <Col xs={3}>
                             <Button