+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix open request count on item when creating a new request. Fixes UIREQ-343.
 * Show effective call number in details pane. Refs UIREQ-366.
 * Fix sorting by request date. Fixes UIREQ-399.
+* Support new versions of mod-inventory and mod-circulation. Refs UIREQ-405.
 
 ## [1.14.0](https://github.com/folio-org/ui-requests/tree/v1.14.0) (2019-12-05)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v1.13.0...v1.14.0)